+++ Classes/GOMClientDelegate.h
@@ -18,7 +18,6 @@
 
 @optional
 
-- (BOOL)gomClientShouldReconnect:(GOMClient *)gomClient;
 - (BOOL)gomClient:(GOMClient *)gomClient shouldReRegisterObserverWithBinding:(GOMBinding *)binding;
 
 @end